+++ app/src/main/java/com/example/pac_man/Point.kt
@@ -4,7 +4,7 @@ import android.content.res.Resources
 import android.graphics.Canvas
 import android.graphics.Paint
 
-abstract class Point(private val resources: Resources, val caseWidth: Int, val caseHeight: Int) {
+abstract class Point(private val resources: Resources, val caseWidth: Float, val caseHeight: Float) {
     protected val paint = Paint()
 
     abstract fun draw(canvas: Canvas, map: Array<IntArray>)
